\name{dupage99}
\alias{dupage99}
\docType{data}
\title{DuPage County assessment ratio data set}
\description{
A random draw of 2000 assessment ratios from DuPage County, IL in 1999.  Sales took place in 1999; the assessments were in place in 1998.  
Statutory assessment rates in DuPage County were 0.33.
}
\usage{data(dupage99)}
\format{
  A data frame with 2000 observations on the following 2 variables.
  \describe{
    \item{\code{av}}{the first column. The assessed value.}
    \item{\code{price}}{the second column. The sales price.}
  }
}

\source{Daniel McMillen.  Data were provided originally by the Illinois Department of Revenue.}


\keyword{Datasets}
