\name{McSpatial-package}
\alias{McSpatial-package}
\alias{McSpatial}
\docType{package}
\title{
Nonparametric spatial data analysis
}
\description{
Locally weighted regression, semiparametric and conditionally parametric regression, fourier and cubic spline functions, 
GMM and linearized spatial logit and probit, k-density functions and counterfactuals, subcenter identification, misclassification probit, 
nonparametric quantile regression and conditional density functions, Machado-Mata decomposition for quantile regressions, spatial AR model, 
repeat sales models, conditionally parametric logit and probit
}
\details{
\tabular{ll}{
Package: \tab McSpatial\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2012-1-29\cr
License: \tab GPL (version 2 or later)\cr
LazyLoad: \tab yes\cr
}
}
\author{
Daniel McMillen

Maintainer: Daniel McMillen <mcmillen@illinois.edu>
}

